requirements.txt
fastapi
uvicorn
pymongo
motor
pytest
pytest-asyncio
python-dotenv
httpx
flake8
